refactor: company profile page styles
