+++ client/src/components/ImageInput/ImageInput.tsx
@@ -1,40 +1,30 @@
 import { AddAPhoto } from "@mui/icons-material";
-import { Button, styled } from "@mui/material";
-
-const VisuallyHiddenInput = styled("input")({
-  clip: "rect(0 0 0 0)",
-  clipPath: "inset(50%)",
-  height: 1,
-  overflow: "hidden",
-  position: "absolute",
-  bottom: 0,
-  left: 0,
-  whiteSpace: "nowrap",
-  width: 1,
-});
 
 type ImageInputProps = {
   setFile: React.Dispatch<React.SetStateAction<File | null>>;
+  key?: string;
 };
 
-export default function ImageInput({ setFile }: ImageInputProps) {
+export default function ImageInput({ setFile, key }: ImageInputProps) {
   return (
-    <>
-      <Button
-        component="label"
-        variant="contained"
-        startIcon={<AddAPhoto />}
-        className="text-center btn btn-info"
+    <div>
+      <label
+        htmlFor={`input-file${key}`}
+        className="cursor-pointer underline text-info flex items-center gap-1"
       >
-        Mudar imagem
-        <VisuallyHiddenInput
-          type="file"
-          accept="image/*"
-          onChange={(e) => {
-            setFile(e.target.files![0]);
-          }}
-        />
-      </Button>
-    </>
+        <span>Adicionar foto</span>
+        <AddAPhoto />
+      </label>
+      <input
+        id={`input-file${key}`}
+        name={`input-file${key}`}
+        className="opacity-0 absolute -z-10"
+        type="file"
+        accept="image/*"
+        onChange={(e) => {
+          setFile(e.target.files![0]);
+        }}
+      />
+    </div>
   );
 }
