+++ client/src/constants/api-routes.ts
@@ -1,5 +1,6 @@
 export const STUDENTS_PATH = "/students";
 export const COMPANIES_PATH = "/companies";
+export const COMPANY_PROFILE_PICTURE_PATH = "/companies/profile/image";
 
 export const PROFILE_STUDENT_PATH = "/students/profile";
 export const PROFILE_COMPANY_PATH = "/companies/profile";
